# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Bank Server API",
    "description": "Production-grade banking backend service API",
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "User login",
        "description": "Authenticate user with email and password",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" },
                "password": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/v1/accounts": {
      "get": {
        "tags": ["accounts"],
        "summary": "List user accounts",
        "security": [{ "BearerAuth": [] }],
        "produces": ["application/json"],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/v1/accounts/balances": {
      "get": {
        "tags": ["accounts"],
        "summary": "List all account balances",
        "security": [{ "BearerAuth": [] }],
        "produces": ["application/json"],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/v1/accounts/{accountId}/balance": {
      "get": {
        "tags": ["accounts"],
        "summary": "Get account balance",
        "security": [{ "BearerAuth": [] }],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not Found" }
        }
      }
    },
    "/api/v1/beneficiaries": {
      "get": {
        "tags": ["beneficiaries"],
        "summary": "List beneficiaries",
        "security": [{ "BearerAuth": [] }],
        "produces": ["application/json"],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/v1/transfer-modes": {
      "get": {
        "tags": ["transfer-modes"],
        "summary": "List transfer modes",
        "security": [{ "BearerAuth": [] }],
        "produces": ["application/json"],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}